+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Fixed
+
+- PR [#169](https://github.com/marinasundstrom/CheckedExceptions/pull/169) Handle local func and lambdas in redundancy check
+
 ## [1.7.0] - 2025-07-30
 
 ### Added